refactor(pages): redeclare some templates as pages and fix kontakt page type errors
