+++ src/pages/lage.tsx
@@ -3,67 +3,69 @@
 // Contains all the functionality necessary to define React components
 //> MDB
 // "Material Design for Bootstrap" is a great UI design framework
+import {fields} from '@snek-at/jaen-pages'
 
 //> Components
-import ImageTopSection from '../../components/organisms/ImageTopSection'
-import {Footer, Navbar} from '../../components/molecules'
+import ImageTopSection from '../components/organisms/ImageTopSection'
+import {Footer, Navbar} from '../components/molecules'
 
 //> CSS
-import * as style from './style'
+import {JaenTemplate} from '@snek-at/jaen-pages/src/types'
+import {Box} from '@chakra-ui/layout'
 
-import {fields, JaenTemplate} from '@snek-at/jaen-pages'
-import {Box} from '@chakra-ui/react'
+import {ResponsiveImage} from '../components/layout/ResponsiveImage'
 //#endregion
 
 //#region > Components
-const ImmoPage: JaenTemplate = () => {
+const LocationPage: JaenTemplate = () => {
   return (
-    <style.responsiveImage>
+    <ResponsiveImage>
       <Box
         as="section"
-        id="immopage"
+        id="locationpage"
+        paddingBottom="12vh"
         overflow="hidden"
         paddingTop="15vh"
-        paddingBottom="12vh"
-        minH="100vh">
+        minHeight="100vh">
         <Navbar />
         <ImageTopSection
           head1={
             <fields.TextField
-              fieldName="immohead1"
+              fieldName="locationhead"
               initValue="Überschrift"
               rtf={false}
             />
           }
           img1={
             <fields.ImageField
-              fieldName="immoimg"
+              fieldName="locationimage"
               initValue={{
                 src: 'https://i.ibb.co/J2jzkBx/placeholder.jpg',
-                title: 'immoimg',
-                alt: 'immoimg'
+                title: 'locationimg',
+                alt: 'locationimg'
               }}
               className="responsiveImage"
             />
           }
           text1={
             <fields.TextField
-              fieldName="immotext1"
+              fieldName="locationrichtext"
               initValue="<p>Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.</p>"
             />
           }
         />
+
         <Footer />
       </Box>
-    </style.responsiveImage>
+    </ResponsiveImage>
   )
 }
 
-ImmoPage.TemplateName = 'ImmoPage'
+LocationPage.TemplateName = 'LocationPage'
 //#endregion
 
 //#region > Exports
-export default ImmoPage
+export default LocationPage
 //#endregion
 
 /**
