+++ src/pages/projekt.tsx
@@ -3,69 +3,68 @@
 // Contains all the functionality necessary to define React components
 //> MDB
 // "Material Design for Bootstrap" is a great UI design framework
-import {fields} from '@snek-at/jaen-pages'
 
 //> Components
-import ImageTopSection from '../../components/organisms/ImageTopSection'
-import {Footer, Navbar} from '../../components/molecules'
+import ImageTopSection from '../components/organisms/ImageTopSection'
+import {Footer, Navbar} from '../components/molecules'
 
 //> CSS
-import {JaenTemplate} from '@snek-at/jaen-pages/src/types'
-import {Box} from '@chakra-ui/layout'
 
-import * as style from './style'
+import {fields, JaenTemplate} from '@snek-at/jaen-pages'
+import {Box} from '@chakra-ui/react'
+import {ResponsiveImage} from '../components/layout/ResponsiveImage'
+
 //#endregion
 
 //#region > Components
-const LocationPage: JaenTemplate = () => {
+const ImmoPage: JaenTemplate = () => {
   return (
-    <style.responsiveImage>
+    <ResponsiveImage>
       <Box
         as="section"
-        id="locationpage"
-        paddingBottom="12vh"
+        id="immopage"
         overflow="hidden"
         paddingTop="15vh"
-        minHeight="100vh">
+        paddingBottom="12vh"
+        minH="100vh">
         <Navbar />
         <ImageTopSection
           head1={
             <fields.TextField
-              fieldName="locationhead"
+              fieldName="immohead1"
               initValue="Überschrift"
               rtf={false}
             />
           }
           img1={
             <fields.ImageField
-              fieldName="locationimage"
+              fieldName="immoimg"
               initValue={{
                 src: 'https://i.ibb.co/J2jzkBx/placeholder.jpg',
-                title: 'locationimg',
-                alt: 'locationimg'
+                title: 'immoimg',
+                alt: 'immoimg'
               }}
               className="responsiveImage"
             />
           }
           text1={
             <fields.TextField
-              fieldName="locationrichtext"
+              fieldName="immotext1"
               initValue="<p>Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.</p>"
             />
           }
         />
-
         <Footer />
       </Box>
-    </style.responsiveImage>
+    </ResponsiveImage>
   )
 }
 
-LocationPage.TemplateName = 'LocationPage'
+ImmoPage.TemplateName = 'ImmoPage'
 //#endregion
 
 //#region > Exports
-export default LocationPage
+export default ImmoPage
 //#endregion
 
 /**
